Teacher Command and Vplan in Table style
Code Changes
- Added Teacher Command - usage in Bot-Help
- Added a Discord Table Library by @Konstantin Kuntzsch for the Vplan.
 -> Vplan now in a table by using Discord `YOUR TEXT` format
- Moved PermissionError.java from de.k7Bot.manage to de.K7Bot.util
- Moved SyntaxError.java.java from de.k7Bot.manage to de.K7Bot.util
- Moved LiteSQL.java.java from de.k7Bot.manage to de.K7Bot.util

+++ src/de/k7bot/music/commands/UebersteuerungAdmin.java
@@ -3,8 +3,8 @@ package de.k7bot.music.commands;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
 import de.k7bot.Klassenserver7bbot;
 import de.k7bot.commands.types.ServerCommand;
-import de.k7bot.manage.SyntaxError;
 import de.k7bot.music.MusicController;
+import de.k7bot.util.SyntaxError;
 
 import java.time.OffsetDateTime;
 import java.util.concurrent.TimeUnit;
@@ -21,14 +21,14 @@ public class UebersteuerungAdmin implements ServerCommand {
 			message.delete().queue();
 			String[] args = message.getContentDisplay().split(" ");
 			try {
-				Integer volume = Integer.valueOf(Integer.parseInt(args[1]));
+				int volume = Integer.parseInt(args[1]);
 				Klassenserver7bbot.INSTANCE.getMusicUtil().updateChannel(channel);
 				if (args.length > 1) {
 					Guild guild = channel.getGuild();
 					MusicController controller = Klassenserver7bbot.INSTANCE.playerManager
 							.getController(guild.getIdLong());
 					AudioPlayer player = controller.getPlayer();
-					player.setVolume(volume.intValue());
+					player.setVolume(volume);
 					Klassenserver7bbot.INSTANCE.getDB()
 							.onUpdate("UPDATE botutil SET volume = " + volume + " WHERE id = " + '\001');
 					EmbedBuilder builder = new EmbedBuilder();
